+++ packages/gui/src/components/html/editor.tsx
@@ -55,8 +55,10 @@ export function HtmlEditor({ onChange }: HtmlEditorProps) {
           borderRightStyle: 'solid',
           borderBottomWidth: '1px',
           borderBottomStyle: 'solid',
+          width: '320px',
           overflowX: 'auto',
           resize: 'horizontal',
+          minHeight: '100svh',
         }}
       >
         <TreeNode
@@ -98,11 +100,13 @@ function NodeEditor({ value, onChange, onRemove }: TagEditorProps) {
         resize: 'horizontal',
         overflowX: 'auto',
         p: 3,
+        width: '320px',
         borderRightWidth: '1px',
         borderRightStyle: 'solid',
         borderBottomWidth: '1px',
         borderBottomStyle: 'solid',
         borderColor: 'border',
+        minHeight: '100svh',
       }}
     >
       <div sx={{ mb: 2, display: 'flex', alignItems: 'flex-end' }}>
@@ -208,7 +212,7 @@ function TreeNode({ value, path, onSelect, onChange }: TreeNodeProps) {
 
   if (typeof value === 'string') {
     return (
-      <div>
+      <div sx={{ cursor: 'default' }}>
         <button
           sx={{
             border: 'none',
@@ -252,8 +256,9 @@ function TreeNode({ value, path, onSelect, onChange }: TreeNodeProps) {
           border: 'none',
           backgroundColor: 'background',
           color: 'text',
+          cursor: 'default',
           ':before': {
-            cursor: 'pointer',
+            cursor: 'default',
             content: open ? '"▼"' : '"▶︎"',
             display: 'inline-block',
             width: '1rem',
@@ -307,7 +312,7 @@ function AddChildButton({ onClick }: { onClick(): void }) {
     <button
       onClick={onClick}
       sx={{
-        cursor: 'pointer',
+        cursor: 'default',
         display: 'block',
         background: 'none',
         border: 'none',